requirements.txt
# Core dependencies
notion-client>=2.2.15
linear-sdk>=1.0.0
gitpython>=3.1.43
pyyaml>=6.0.1
python-dotenv>=1.0.1
requests>=2.31.0
aiohttp>=3.9.5

# Webhook server
flask>=3.0.3
waitress>=3.0.0

# CLI and UI
rich>=13.7.1
click>=8.1.7
tqdm>=4.66.4

# Data processing
pydantic>=2.7.0
pandas>=2.2.2

# MCP server (optional)
mcp>=0.1.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.6
pytest-cov>=5.0.0

# Linting and formatting
ruff>=0.4.4
black>=24.4.2
mypy>=1.10.0

# Development
ipython>=8.24.0
watchdog>=4.0.0
